[bugfix] Rename RefreshButton in selection toolbox to avoid naming conflict
Renamed graph/selectionToolbox/RefreshButton.vue to RefreshSelectionButton.vue
to resolve component naming conflict with common/RefreshButton.vue.
