+++ src/components/graph/SelectionToolbox.vue
@@ -15,7 +15,7 @@
     <MaskEditorButton />
     <ConvertToSubgraphButton />
     <DeleteButton />
-    <RefreshButton />
+    <RefreshSelectionButton />
     <ExtensionCommandButton
       v-for="command in extensionToolboxCommands"
       :key="command.id"
@@ -39,7 +39,7 @@ import ExtensionCommandButton from '@/components/graph/selectionToolbox/Extensio
 import HelpButton from '@/components/graph/selectionToolbox/HelpButton.vue'
 import MaskEditorButton from '@/components/graph/selectionToolbox/MaskEditorButton.vue'
 import PinButton from '@/components/graph/selectionToolbox/PinButton.vue'
-import RefreshButton from '@/components/graph/selectionToolbox/RefreshButton.vue'
+import RefreshSelectionButton from '@/components/graph/selectionToolbox/RefreshSelectionButton.vue'
 import { useCanvasInteractions } from '@/composables/graph/useCanvasInteractions'
 import { useExtensionService } from '@/services/extensionService'
 import { type ComfyCommandImpl, useCommandStore } from '@/stores/commandStore'
