+++ src/components/graph/selectionToolbox/RefreshSelectionButton.vue
